refactor: streamline SEO component and move embed generation logic to a separate module
